fixed version of drunken delivery service app

+++ src/main/java/com/example/springblog/repo/Users.java
@@ -3,7 +3,7 @@ package com.example.springblog.repo;
 import com.example.springblog.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface Users extends CrudRepository<User, Integer> {
+public interface Users extends CrudRepository<User, Long> {
         User findById(long id);
         User findByUsername(String username);
         User findByEmail(String email);
